added MathGame.Data
