+++ src/MathGame.ConsoleApplication/Utilities/Helpers.cs
@@ -10,26 +10,9 @@ using MathGame.Models;
 namespace MathGame.ConsoleApplication.Utilities;
 internal class Helpers
 {
-
     #region Methods: Internal
 
-    internal static List<Game> games = new List<Game>
-    {
-        //Providing Sample Game History Data
-        new Game { Date = DateTime.Now.AddDays(1), Type = GameType.Addition, Score = 5 },
-        new Game { Date = DateTime.Now.AddDays(2), Type = GameType.Mulitplication, Score = 4 },
-        new Game { Date = DateTime.Now.AddDays(3), Type = GameType.Division, Score = 4 },
-        new Game { Date = DateTime.Now.AddDays(4), Type = GameType.Subtraction, Score = 3 },
-        new Game { Date = DateTime.Now.AddDays(5), Type = GameType.Addition, Score = 1 },
-        new Game { Date = DateTime.Now.AddDays(6), Type = GameType.Mulitplication, Score = 2 },
-        new Game { Date = DateTime.Now.AddDays(7), Type = GameType.Division, Score = 3 },
-        new Game { Date = DateTime.Now.AddDays(8), Type = GameType.Subtraction, Score = 4 },
-        new Game { Date = DateTime.Now.AddDays(9), Type = GameType.Addition, Score = 4 },
-        new Game { Date = DateTime.Now.AddDays(10), Type = GameType.Mulitplication, Score = 1 },
-        new Game { Date = DateTime.Now.AddDays(11), Type = GameType.Subtraction, Score = 0 },
-        new Game { Date = DateTime.Now.AddDays(12), Type = GameType.Division, Score = 2 },
-        new Game { Date = DateTime.Now.AddDays(13), Type = GameType.Subtraction, Score = 5 },
-    };
+    
 
     internal static int[] GetDivisionNumbers()
     {
@@ -51,34 +34,8 @@ internal class Helpers
         return result;
     }
 
-    internal static void AddToHistory(int gameScore, GameType gameType)
-    {
-        games.Add(new Game
-        {
-            Date = DateTime.Now,
-            Score = gameScore,
-            Type = gameType
-        });
-    }
+   
 
-    internal static void PrintGames()
-    {
-        // var gamesToPrint = games.Where(x => x.Type == GameType.Division);
-        // var gamesToPRint = games.Where(x => x.Date > new DateTime(2022, 08, 09));
-        // var gamesToPRint = games.Where(x => x.Date > new DateTime(2022, 08, 09) && x.Score > 3);
-        var gamesToPrint = games.Where(x => x.Date > new DateTime(2022, 08, 09)).OrderByDescending(x => x.Score);
-
-        Console.Clear();
-        Console.WriteLine("Games History");
-        Console.WriteLine("---------------------------");
-        foreach (var game in gamesToPrint)
-        {
-            Console.WriteLine($"{game.Date} - {game.Type}: {game.Score}pts");
-        }
-        Console.WriteLine("---------------------------\n");
-        Console.WriteLine("Press any key to return to Main Menu");
-        Console.ReadLine();
-    }
 
     internal static string? ValidateResult(string result)
     {
