+++ src/MathGame.Data/MathGame.Data.csproj
@@ -1,0 +1,15 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net8.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\MathGame.ConsoleApplication\MathGame.ConsoleApplication.csproj" />
+    <ProjectReference Include="..\MathGame\MathGame.csproj" />
+  </ItemGroup>
+
+</Project>
