+++ src/MathGame.sln
@@ -5,7 +5,9 @@ VisualStudioVersion = 17.8.34330.188
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MathGame", "MathGame\MathGame.csproj", "{1958C166-27AA-4FE6-A735-30227C699F89}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MathGame.ConsoleApplication", "MathGame.ConsoleApplication\MathGame.ConsoleApplication.csproj", "{4914E134-125A-403F-BAF0-5F9CEB94F412}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MathGame.ConsoleApplication", "MathGame.ConsoleApplication\MathGame.ConsoleApplication.csproj", "{4914E134-125A-403F-BAF0-5F9CEB94F412}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MathGame.Data", "MathGame.Data\MathGame.Data.csproj", "{FA675C42-A26C-451A-94F8-4C624524D15F}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,6 +23,10 @@ Global
 		{4914E134-125A-403F-BAF0-5F9CEB94F412}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4914E134-125A-403F-BAF0-5F9CEB94F412}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{4914E134-125A-403F-BAF0-5F9CEB94F412}.Release|Any CPU.Build.0 = Release|Any CPU
+		{FA675C42-A26C-451A-94F8-4C624524D15F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FA675C42-A26C-451A-94F8-4C624524D15F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FA675C42-A26C-451A-94F8-4C624524D15F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FA675C42-A26C-451A-94F8-4C624524D15F}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
