+++ src/MathGame/Enums/GameDifficulty.cs
@@ -1,7 +1,13 @@
-﻿namespace MathGame.Enums;
+﻿// -------------------------------------------------------------------------------------------------
+// MathGame.Enums.GameType
+// -------------------------------------------------------------------------------------------------
+// Enum for the game difficulty.
+// -------------------------------------------------------------------------------------------------
+
+namespace MathGame.Enums;
 public enum GameDifficulty
 {  
-    Easy,
-    Normal,
-    Hard
+    Easy = 1,
+    Normal = 2,
+    Hard = 3
 }
