+++ src/MathGame/Models/Game.cs
@@ -12,7 +12,6 @@ public class Game
     public DateTime Date {  get; set; }
     public int Score { get; set; }
     public GameType Type { get; set; }
-    public GameDifficulty Difficulty { get; set; }
 
     #endregion
 }
